Refactor PBR shader system and improve light handling
Updated shader variable names and locations for consistency and clarity across C# and GLSL code. Enhanced light handling in the engine, including improved support for light types, range, and selection visualization. Fixed float3 multiplication bug, improved transform axis handling, and made temp directory deletion safer. Updated project output type and improved PBR material parameter handling for models.

+++ scythe/path.cs
@@ -42,7 +42,8 @@ public abstract class path {
     }
 
     public static void clear_temp() {
-        
-        Directory.Delete(tmp_dir, true);
+
+        if (Directory.Exists(tmp_dir))
+            Directory.Delete(tmp_dir, true);
     }
 }
